Rename TaskConfiguration to MigrationTaskConfiguration
